% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeToolFunctions.R
\name{appendInteraction}
\alias{appendInteraction}
\title{Appends Interactions to a \code{\link[JoeTools:report]{report()}} Output Dataframe}
\usage{
appendInteraction(report, df, Measures, Factor, Interaction)
}
\arguments{
\item{report}{The report dataframe that is being appended}

\item{df}{The orginal dataframe used for the report}

\item{Measures}{A vector of string names for the dependant variable. Order must be the same as the repor}

\item{Factor}{A binary factor (as a string) which will be used for comparisons.  Must be 1 (Treatment) or 0 (No_Treatment)}

\item{Interaction}{A factor column name (as a string) that will be used as an interacting independant variable.}
}
\value{

}
\description{
Appends Interactions to a \code{\link[JoeTools:report]{report()}} Output Dataframe
}
