% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeToolFunctions.R
\name{CSTable}
\alias{CSTable}
\title{Simplifies Campus Solutions Table Calls}
\usage{
CSTable(TABLENAME)
}
\arguments{
\item{TABLENAME}{Name of the table. Converted automatically to UPPERCASE}
}
\value{

}
\description{
Simplifies Campus Solutions Table Calls
}
