% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeToolFunctions.R
\name{document}
\alias{document}
\title{Creates a document including the Report and Pairwise comparisons produced above}
\usage{
document(Report, Pairwise, measureNames, factorNames, percentOutcomes)
}
\arguments{
\item{Report}{The report output by report/appendInteraction}

\item{Pairwise}{The output of the pairwise function created from Report}

\item{measureNames}{A named vector with cleaned names for Measures}

\item{factorNames}{A named vector with cleaned names for Factors}

\item{percentOutcomes}{A vector of Outcomes that should be percentages}
}
\value{

}
\description{
Creates a document including the Report and Pairwise comparisons produced above
}
