% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeToolFunctions.R
\name{edwFinance}
\alias{edwFinance}
\title{Pulling Finance EDW Tables}
\usage{
edwFinance(TableName, DB = EDW)
}
\arguments{
\item{TableName}{String. Name of the Table to be pulled.}

\item{DB}{Reference to the connection. Defaults to EDW}
}
\value{

}
\description{
Pulling Finance EDW Tables
}
