% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeToolFunctions.R
\name{FitFlextableToPage}
\alias{FitFlextableToPage}
\title{Force Flextable to Page}
\usage{
FitFlextableToPage(ft, pgwidth = 6)
}
\arguments{
\item{ft}{Flextable object}

\item{pgwidth}{Customization page width setting}
}
\value{

}
\description{
Force Flextable to Page
}
