% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeToolFunctions.R
\name{multiplot}
\alias{multiplot}
\title{Plotting more than one plot together}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{Name of each plot}

\item{plotlist}{A vector with plot names}

\item{cols}{How many columns of plots you want in your chart}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\value{

}
\description{
Plotting more than one plot together
}
