% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeToolFunctions.R
\name{pairwiseReport}
\alias{pairwiseReport}
\title{Produce the Pairwise Comparisons found Significant in a Report}
\usage{
pairwiseReport(Report, Data)
}
\arguments{
\item{Report}{Report object output from Report/appendInteraction}

\item{Data}{The original data used to make Report}
}
\value{

}
\description{
Produce the Pairwise Comparisons found Significant in a Report
}
