% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeToolFunctions.R
\name{report}
\alias{report}
\title{Creates a Report with t-tests for a Vector of Outcomes}
\usage{
report(df, Measures, Factor, paired = c("Yes", "No", "Try"))
}
\arguments{
\item{df}{Dataframe to use for the report}

\item{Measures}{A vector of string names for columns to take means for. Order will be used to make the report.}

\item{Factor}{A binary factor which will be used for comparisons.  Must be 1 (Treatment) or 0 (No_Treatment)}

\item{paired}{Options for running t-test.  "Yes" forces paired, t-tests, "No" assumes no pairing, and "Try" will try a paired t-test, and follow up with a unpaired t-test if it fails to run.}
}
\value{

}
\description{
Creates a Report with t-tests for a Vector of Outcomes
}
