% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeToolFunctions.R
\name{se}
\alias{se}
\title{Standard Error}
\usage{
se(data)
}
\arguments{
\item{data}{The vector or dataset over which the standard error is calculated}
}
\value{
The calculated Standard Error
}
\description{
Standard Error
}
\examples{
x=c(10, 20, 30, 40, 50)
se(x)
}
