% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeToolFunctions.R
\name{simplifyAppend}
\alias{simplifyAppend}
\title{Simplifying Reports}
\usage{
simplifyAppend(input)
}
\arguments{
\item{input}{dataframe to simplify results from}
}
\value{
Simplified interacton reports
}
\description{
Simplifying Reports
}
