% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeToolFunctions.R
\name{SISTable}
\alias{SISTable}
\title{Simplifies SIS Table Calls}
\usage{
SISTable(TABLENAME)
}
\arguments{
\item{TABLENAME}{Name of the table. Converted automatically to UPPERCASE}
}
\value{

}
\description{
Simplifies SIS Table Calls
}
