% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeToolFunctions.R
\name{wbsave}
\alias{wbsave}
\title{Flexible Workbook Creation Function}
\usage{
wbsave(df, filename, sheetBy = NULL, keepNames = FALSE)
}
\arguments{
\item{df}{The dataframe to be saved}

\item{filename}{Name of the file, including the xlsx extension}

\item{sheetBy}{Option to create sheets based on the values in a particular column of the dataframe}

\item{keepNames}{The function automatically renames the column names with \code{\link[stringr:case]{stringr::str_to_sentence()}}. Setting this to TRUE will leave your column names unaltered}
}
\value{
Creates a Excel workbook from the supplied dataframe
}
\description{
Flexible Workbook Creation Function
}
\examples{
library(datasets)
library(tidyverse)
library(xlsx)
data(iris)
wbsave(iris, "Iris report.xlsx")
wbsave(iris, "Iris report - by Species.xlsx", sheetBy="Species", keepNames=TRUE)
}
