% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoeToolFunctions.R
\name{pairwise}
\alias{pairwise}
\title{Produce the Pairwise Comparisons on factors in a group}
\usage{
pairwise(df, Ind, Dep, Factor, Logit = FALSE)
}
\arguments{
\item{df}{Dataframe with by subject Factor, Independent and Dependent Variable info}

\item{Ind}{Column with the independant variable}

\item{Dep}{Column with the dependant variable}

\item{Factor}{Column over which to create comparisons}

\item{Logit}{Allows for logistic regression if Logit=TRUE, off by default}
}
\value{

}
\description{
Produce the Pairwise Comparisons on factors in a group
}
